requirements.txt
# Core Django
Django==5.2.5
djangorestframework==3.15.2
djoser==2.2.3

# Database
psycopg2-binary==2.9.9
django-redis==5.4.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.4.0
python-decouple==3.8

# File handling
Pillow==10.4.0

# Production server
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring & Logging
sentry-sdk[django]==1.38.0

# Background tasks (optional)
celery==5.3.4
redis==5.0.1

# Development & Testing
pytest==7.4.3
pytest-django==4.7.0
factory-boy==3.3.0
coverage==7.3.2

# Code quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Documentation
drf-spectacular==0.27.0